migrate user-dashboard into default tabbed view
